updated metadata function
The previous metadata function was not passing useful location information, had naming errors and was writing over itself. The new function now saves the x and y pixel location of the top left corner of the tile relative to the SAR image. Additionally, it checks if a metadata file already exists and creates one if not or otherwise updates.